test: merge ThinkBlock.test.ts into ThinkBlock.spec.ts
Deduplicates the two ThinkBlock test files by absorbing the 3 unique
markdown-rendering and max-height tests into the main spec.
